+++ Model/Command/AbstractCommand.php
@@ -21,7 +21,6 @@ use Magento\Payment\Model\InfoInterface;
 use Magento\Sales\Model\Order as SalesOrder;
 use Magento\Sales\Model\Order\Payment as OrderPayment;
 use Psr\Log\LoggerInterface;
-
 use function get_class;
 
 /**
@@ -149,7 +148,7 @@ abstract class AbstractCommand implements CommandInterface
         $customer = $this->_getClient()->fetchCustomer($customerId);
 
         if (!$this->_orderHelper->validateGatewayCustomerAgainstOrder($order, $customer)) {
-            throw new LocalizedException(__('Payment information does not match billing address.'));
+            $this->_orderHelper->updateGatewayCustomerFromOrder($order, $customer);
         }
 
         return $customerId;