+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [1.0.2][1.0.2]
+### Fix
+* Billing address not changeable in checkout.
+* Checkout not possible with virtual quotes.
+
 ## [1.0.1][1.0.1]
 ### Added
 * php 7.4 support.
@@ -15,3 +20,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 [1.0.0]: https://github.com/heidelpay/magento2-merchant-gateway/tree/1.0.0
 [1.0.1]: https://github.com/heidelpay/magento2-merchant-gateway/compare/1.0.0..1.0.1
+[1.0.2]: https://github.com/heidelpay/magento2-merchant-gateway/compare/1.0.1..1.0.2